Split up shared reducer into shared list reducer and shared entry reducer respectively
Shared entry reducers handler single entries, and has a different initial state than the list reducers.
