+++ src/sentry/static/sentry/app/redux/reducers/sharedList.js
@@ -114,59 +114,6 @@ export function createEntryFailure(state, action) {
   };
 }
 
-export function updateEntryRequest(state, action) {
-  return {
-    ...state,
-    updating: true,
-    errorMessage: null,
-  };
-}
-
-export function updateEntrySuccess(state, action) {
-  const byIds = {};
-  byIds[action.entry.id] = action.entry;
-  return {
-    ...state,
-    updating: false,
-    byIds: merge({}, state.byIds, byIds),
-  };
-}
-
-export function updateEntryFailure(state, action) {
-  return {
-    ...state,
-    updating: false,
-    errorMessage: action.message,
-  };
-}
-
-export function getEntryRequest(state, action) {
-  return {
-    ...state,
-    loadingDetails: true,
-    detailsId: null,
-  };
-}
-
-export function getEntrySuccess(state, action) {
-  const byIds = {};
-  byIds[action.entry.id] = action.entry;
-  return {
-    ...state,
-    loadingDetails: false,
-    detailsId: action.entry.id,
-    byIds: merge({}, state.byIds, byIds),
-  };
-}
-
-export function getEntryFailure(state, action) {
-  return {
-    ...state,
-    errorMessage: action.message,
-    loadingDetails: false,
-  };
-}
-
 const createResourceReducer = (resource, initialState) => (
   state = initialState,
   action
@@ -182,25 +129,12 @@ const createResourceReducer = (resource, initialState) => (
       return selectSingle(state, action);
     case `SELECT_PAGE_OF_${resource}`:
       return selectAll(state, action);
-    case `GET_${resource}_REQUEST`:
-      return getEntryRequest(state, action);
-    case `GET_${resource}_SUCCESS`:
-      return getEntrySuccess(state, action);
-    case `GET_${resource}_FAILURE`:
-      return getEntryFailure(state, action);
     case `CREATE_${resource}_REQUEST`:
       return createEntryRequest(state, action);
     case `CREATE_${resource}_SUCCESS`:
       return createEntrySuccess(state, action);
     case `CREATE_${resource}_FAILURE`:
       return createEntryFailure(state, action);
-    case `UPDATE_${resource}_REQUEST`:
-      return updateEntryRequest(state, action);
-    case `UPDATE_${resource}_SUCCESS`:
-      return updateEntrySuccess(state, action);
-    case `UPDATE_${resource}_FAILURE`:
-      return updateEntryFailure(state, action);
-
     default:
       return state;
   }
@@ -227,6 +161,9 @@ export const list = {
   },
 
   // All reducers available for that list:
+  createEntryRequest,
+  createEntrySuccess,
+  createEntryFailure,
   selectSingle,
   selectAll,
   getListRequest,
@@ -234,32 +171,9 @@ export const list = {
   getListFailure,
 };
 
-export const entry = {
-  // Required state for following an entry protocol
-  initialState: {
-    loadingDetails: false,
-    updating: false,
-    detailsId: null,
-  },
-
-  // All reducers
-  createEntryRequest,
-  createEntrySuccess,
-  createEntryFailure,
-  getEntryRequest,
-  getEntrySuccess,
-  getEntryFailure,
-  updateEntryRequest,
-  updateEntrySuccess,
-  updateEntryFailure,
-};
-
 // A resource follows both the list and entry protocols
 export const resource = {
-  initialState: {
-    ...entry.initialState,
-    ...list.initialState,
-  },
+  initialState: {...list.initialState},
 
   createReducer: createResourceReducer,
 };
