+++ src/sentry/static/sentry/app/redux/reducers/workBatch.js
@@ -1,4 +1,4 @@
-import {resource} from './shared';
+import {resource} from 'app/redux/reducers/sharedList';
 import {RESOURCE_NAME} from '../actions/workBatch.js';
 
 export const initialState = {
